arreglando el problema del backend

+++ backend/src/main/java/com/styp/cenate/api/mbac/PermisosController.java
@@ -1,11 +1,7 @@
 package com.styp.cenate.api.mbac;
-import lombok.Data;
 
-import com.styp.cenate.dto.mbac.CheckPermisoRequestDTO;
-import com.styp.cenate.dto.mbac.CheckPermisoResponseDTO;
-import com.styp.cenate.dto.mbac.ModuloSistemaResponse;
-import com.styp.cenate.dto.mbac.PaginaModuloResponse;
-import com.styp.cenate.dto.mbac.PermisoUsuarioResponseDTO;
+import lombok.Data;
+import com.styp.cenate.dto.mbac.*;
 import com.styp.cenate.service.mbac.PermisosService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +18,7 @@ import java.util.List;
  * y lista módulos o páginas accesibles para el usuario autenticado.
  *
  * Autor: CENATE Development Team
- * Versión: 1.2 (actualizado)
+ * Versión: 1.3 (actualizado con compatibilidad Long/Integer)
  */
 @Slf4j
 @RestController
@@ -59,9 +55,10 @@ public class PermisosController {
     @GetMapping("/usuario/{userId}/modulo/{idModulo}")
     public ResponseEntity<List<PermisoUsuarioResponseDTO>> obtenerPermisosPorUsuarioYModulo(
             @PathVariable Long userId, @PathVariable Integer idModulo) {
+
         log.info("📥 GET /api/mbac/permisos/usuario/{}/modulo/{}", userId, idModulo);
         List<PermisoUsuarioResponseDTO> permisos =
-                permisosService.obtenerPermisosPorUsuarioYModulo(userId, idModulo);
+                permisosService.obtenerPermisosPorUsuarioYModulo(userId, idModulo.longValue());
         return ResponseEntity.ok(permisos);
     }
 
@@ -93,8 +90,10 @@ public class PermisosController {
     @GetMapping("/usuario/{userId}/modulo/{idModulo}/paginas")
     public ResponseEntity<List<PaginaModuloResponse>> obtenerPaginasAccesiblesUsuario(
             @PathVariable Long userId, @PathVariable Integer idModulo) {
+
         log.info("📑 GET /api/mbac/permisos/usuario/{}/modulo/{}/paginas", userId, idModulo);
-        List<PaginaModuloResponse> paginas = permisosService.obtenerPaginasAccesiblesUsuario(userId, idModulo);
+        List<PaginaModuloResponse> paginas =
+                permisosService.obtenerPaginasAccesiblesUsuario(userId, idModulo.longValue());
         return ResponseEntity.ok(paginas);
     }
 }
